src/use-cases/search-pet-by-city.ts
import { PetsRepository } from '@/repositories/pets-repository'
import { Pet } from '@prisma/client'
import { ResourceNotFoundError } from './errors/resource-not-found-error'

interface FetchPetByCityUseCaseRequest {
  query: string
  page: number
}

interface FetchPetByCityUseCaseResponse {
  pet: Pet[]
}

export class FetchPetByCityUseCase {
  constructor(private petsRepository: PetsRepository) {}

  async execute({
    query,
    page,
  }: FetchPetByCityUseCaseRequest): Promise<FetchPetByCityUseCaseResponse> {
    const pet = await this.petsRepository.searchMany(query, page)

    if (!pet) {
      throw new ResourceNotFoundError()
    }

    return {
      pet,
    }
  }
}
